Move the layer tests to their own file.
