+++ tests/UnitTests/CoreGraphics.drawing/CGContextDrawingTests.cpp
@@ -237,73 +237,3 @@ DISABLED_DRAW_TEST_F(CGContext, ChangeCTMAfterCreatingPath, WhiteBackgroundTest)
     CGContextStrokePath(context);
     CGContextRestoreGState(context);
 }
-
-static void _strokeTwoCirclesInContext(CGContextRef context, CGRect bounds) {
-    CGPoint center = _CGRectGetCenter(bounds);
-    CGRect centerEllipseRect = _CGRectCenteredOnPoint({ 150, 150 }, center);
-    CGFloat translations[]{ -60.f, +60.f };
-
-    for (float xSlide : translations) {
-        CGRect translatedRect = CGRectApplyAffineTransform(centerEllipseRect, CGAffineTransformMakeTranslation(xSlide, 0));
-        CGContextStrokeEllipseInRect(context, translatedRect);
-    }
-}
-
-DRAW_TEST_F(CGContext, TransparencyLayerIsShadowedAllAtOnce, WhiteBackgroundTest) {
-    CGContextRef context = GetDrawingContext();
-    CGRect bounds = GetDrawingBounds();
-
-    CGContextSetShadow(context, CGSize{ 10.f, 10.f }, 1.0);
-
-    CGContextBeginTransparencyLayer(context, nullptr);
-
-    CGContextSetRGBFillColor(context, 1.0, 0.0, 0.0, 1.0);
-    CGContextSetRGBStrokeColor(context, 1.0, 0.0, 0.0, 1.0);
-    CGContextSetLineWidth(context, 5);
-
-    CGPoint center = _CGRectGetCenter(bounds);
-
-    CGContextMoveToPoint(context, center.x - 50, center.y);
-    CGContextAddLineToPoint(context, center.x + 50, center.y);
-    CGContextStrokePath(context);
-
-    CGContextMoveToPoint(context, center.x, center.y - 50);
-    CGContextAddLineToPoint(context, center.x, center.y + 50);
-    CGContextStrokePath(context);
-
-    CGContextEndTransparencyLayer(context);
-}
-
-DRAW_TEST_F(CGContext, TransparencyLayerWithInnerShadowCanStack, WhiteBackgroundTest) {
-    CGContextRef context = GetDrawingContext();
-    CGRect bounds = GetDrawingBounds();
-
-    CGContextSetShadow(context, CGSize{ 10.f, 10.f }, 1.0);
-
-    CGContextBeginTransparencyLayer(context, nullptr);
-
-    CGContextSetShadow(context, CGSize{ -10.f, -10.f }, 1.0);
-
-    CGContextSetRGBStrokeColor(context, 1.0, 0.0, 0.0, 1.0);
-    CGContextSetLineWidth(context, 5);
-
-    _strokeTwoCirclesInContext(context, bounds);
-
-    CGContextEndTransparencyLayer(context);
-}
-
-DRAW_TEST_F(CGContext, TransparencyLayerIsAlphaCompositedAllAtOnce, WhiteBackgroundTest) {
-    CGContextRef context = GetDrawingContext();
-    CGRect bounds = GetDrawingBounds();
-
-    CGContextSetAlpha(context, 0.5);
-
-    CGContextBeginTransparencyLayer(context, nullptr);
-
-    CGContextSetRGBStrokeColor(context, 1.0, 0.0, 0.0, 1.0);
-    CGContextSetLineWidth(context, 5);
-
-    _strokeTwoCirclesInContext(context, bounds);
-
-    CGContextEndTransparencyLayer(context);
-}
